fq.py: 	Minor changes, added argument to player
fq_t.py:
	Now checks whether target files exist before
	attempting to view/rename them.

+++ fq_t.py
@@ -70,6 +70,9 @@ def remove_top(f_name: str,f_dir: str) -> str:
 	# f_name and f_dir are just strs
 	# remove the first line from f_dir/f_name. (a file).
 	# returns the line that was removed, if it exists, else None
+	if not os.path.exists(os.path.join(f_dir,f_name)):
+		print("Queue file does not exist.")
+		return None
 	tmp_f = tempfile.mktemp(dir=f_dir) # type: str
 	os.rename(os.path.join(f_dir,f_name),tmp_f)
 	r = None # type: str
@@ -83,6 +86,9 @@ def remove_top(f_name: str,f_dir: str) -> str:
 	return r.rstrip() if r else None
 
 def add_bottom(f_name: str,f_dir: str,line: str) -> None:
+	"""
+	Appends line to file f_name residing in f_dir.
+	"""
 	l = line.rstrip() # type: str
 	if len(l)>0:
 		with open(os.path.join(f_dir,f_name),'a') as f:
@@ -93,7 +99,12 @@ def view_vid(v_dir: str,v_name: str = "1"):
 	# uses global 'player' and 'player_args'
 	# on linux, this blocks the terminal
 	# returns the 'return code'
-	return subprocess.run([player,os.path.join(v_dir,v_name)]+player_args).returncode
+	vid = os.path.join(v_dir,v_name)
+	if os.path.exists(vid):
+		return subprocess.run([player,vid]+player_args).returncode
+	else:
+		print("No video file exists.")
+		return None
 
 def push_out(v_dir: str,v_name: str = "1",v_name_old: str = "0") -> None:
 	# displaces last vid by replacing v_name_Old with it
@@ -109,6 +120,9 @@ def clear_file(tmp_dir,q_name) -> None:
 	open(os.path.join(tmp_dir,q_name),'w').close()
 
 def yt_dl_bin(url: str,v_dir: str,v_name: str ="1"):
+	"""
+	Downloads video with youtube-dl binary.
+	"""
 	speed = get_envvar("speed") # type: str
 	quality = get_envvar("quality") # type: str
 	cmd = [yt_dl,"-f",quality,"-r",speed,url,"-o",os.path.join(v_dir,v_name)]+yt_dl_args # type: List[str]
@@ -116,6 +130,9 @@ def yt_dl_bin(url: str,v_dir: str,v_name: str ="1"):
 	return subprocess.Popen(cmd)
 
 def case_go(args: List[str],resume: bool = False) -> None:
+	"""
+	Downloads next item in queue or supplied url from args.
+	"""
 	url = None # type: str
 	if len(args)>1:
 		url = args[1]
@@ -130,16 +147,30 @@ def case_go(args: List[str],resume: bool = False) -> None:
 	view_vid(tmp_dir)
 	
 def case_default(args: List[str]) -> None:
+	"""
+	Default case.
+	Downloads next item in queue.
+	"""
 	case_go(args,resume=False)
 
 def case_resume(args: List[str]) -> None:
+	"""
+	Resumes download of last video. Requires url to be in args OR as next item in queue.
+	Downloads regardless of wheter the url corresponds to last video.
+	"""
+	print("Resuming video download.")
 	case_go(args,resume=True)
 
 def case_flush(args: List[str]) -> None:
+	"""
+	Clears contents of queue. Also creates it if doesn't already exist.
+	"""
 	clear_file(tmp_dir,q_name)
 	print("Queue was cleared.")
 
 def case_add(args: List[str]) -> None:
+	"""
+	Adds item to queue."""
 	if len(args)>1:
 		url = args[1] # type: str
 		if url:
@@ -149,38 +180,49 @@ def case_add(args: List[str]) -> None:
 		print("URL was blank.")
 
 def case_top(args: List[str]) -> None:
-	url = top(open(os.path.join(tmp_dir,q_name),'r')) # type :str
-	if url:
-		print(url.rstrip())
-	else:
-		print("Queue was empty.")
+	"""
+	Print next item in queue if possible.
+	"""
+	qfile = os.path.join(tmp_dir,q_name)
+	if os.path.exists(qfile):
+		url = top(open(qfile,'r')) # type :str
+		if url:
+			print(url.rstrip())
+			return
+	print("Queue was empty.")
 
 def case_see(args: List[str]) -> None:
-	with open(os.path.join(tmp_dir,q_name),'r') as f:
-		for l in f:
-			print(l,end='')
-			#print(l)
+	"""
+	Print queue contents to stdout.
+	"""
+	qfile = os.path.join(tmp_dir,q_name)
+	if os.path.exists(qfile):
+		with open(qfile,'r') as f:
+			for l in f:
+				print(l,end='')
 
 def case_view(args: List[str]) -> None:
+	"""
+	Launch viewer and player last downloaded video.
+	"""
 	print("Launching {:s}..".format(player))
 	view_vid(tmp_dir)
 
 ops = {	
 #keyword function req_args
 	"view": case_view,
-	"see" : case_see, #0
-	"top" : case_top, #0
-	"fls" : case_flush, #0
-	"resm": case_resume, #0
-	"*"   : case_default, #0
-	#"int": case_internal, #0
+	"see" : case_see,
+	"top" : case_top,
+	"fls" : case_flush,
+	"resm": case_resume,
+	"*"   : case_default,
 	"go"  : case_go, #1
 	"add" : case_add, #1
+	#"int": case_internal,
 } # type: Dict[str,Callable]
 
-if len(sys.argv)<2:
+if len(sys.argv)<2 or not sys.argv[1] in ops:
 	ops["*"]([])
 else:
 	args = sys.argv[1:] # type: List[str]
-	if args[0] in ops:
-		ops[args[0]](args)
+	ops[args[0]](args)
